MidiParser::ParseAll 100%, DataArraySongInfo frsp fixes
- MidiParser::ParseAll: swap `which`/`startTick` declaration order
  so `which` (initialized to -1) claims r29 before startTick, fixing
  r29↔r30 register swap across the entire loop body (100% match)
- MidiParser::OnSecOffset: add fp_contract off pragma, inline f3 temp
- DataArraySongInfo ctor: add (float) casts to mMuteVolume and
  mVocalMuteVolume assignments to generate correct frsp instructions

+++ src/system/meta/DataArraySongInfo.cpp
@@ -25,10 +25,10 @@ DataArraySongInfo::DataArraySongInfo(
         mHopoThreshold = member_arr->Int(1);
     }
     if (FIND_WITH_BACKUP(mute_volume)) {
-        mMuteVolume = member_arr->Float(1);
+        mMuteVolume = (float)member_arr->Float(1);
     }
     if (FIND_WITH_BACKUP(mute_volume_vocals)) {
-        mVocalMuteVolume = member_arr->Float(1);
+        mVocalMuteVolume = (float)member_arr->Float(1);
     }
     if (FIND_WITH_BACKUP(pans)) {
         DataArray *pan_arr = member_arr->Array(1);
